- got rid of all the std namespace usage in Box2D. - reverted all includes to the c-style, e.g. limits.h - changed the Box2D Visual Studio warning level to 4 (and fixed any warnings).
Former-commit-id: 2adf83cfd79f6ec86e95f4c0aa15252c3150fadf

+++ Box2D/Box2D/Collision/b2Collision.cpp
@@ -220,7 +220,7 @@ int32 b2ClipSegmentToLine(b2ClipVertex vOut[2], const b2ClipVertex vIn[2],
 		vOut[numOut].v = vIn[0].v + interp * (vIn[1].v - vIn[0].v);
 
 		// VertexA is hitting edgeB.
-		vOut[numOut].id.cf.indexA = vertexIndexA;
+		vOut[numOut].id.cf.indexA = static_cast<uint8>(vertexIndexA);
 		vOut[numOut].id.cf.indexB = vIn[0].id.cf.indexB;
 		vOut[numOut].id.cf.typeA = b2ContactFeature::e_vertex;
 		vOut[numOut].id.cf.typeB = b2ContactFeature::e_face;
